requirements.txt
defusedxml~=0.7.1
xmltodict~=0.13.0
libvirt-python~=9.4.0
prometheus-client~=0.17.0